Update for lein 2 and pallet 0.8
Update for lein 2 error handling.  Adds the use of a :pallet profile if
available, with fallback on and internal profile.  This allows the plugin to
work outside of a project, and still allows a project to use different pallet
dependencies if required.

This will only function with a version of pallet higher than 0.8.0-alpha.8.